Cleanup and removal of redundant code by adding utility class.
