+++ src/main/java/edu/harvard/h2ms/service/ReportWorkerComplianceWarning.java
@@ -1,22 +1,17 @@
 package edu.harvard.h2ms.service;
 
-import com.opencsv.CSVWriter;
 import edu.harvard.h2ms.domain.core.Event;
 import edu.harvard.h2ms.domain.core.Question;
 import edu.harvard.h2ms.domain.core.User;
 import edu.harvard.h2ms.exception.InvalidAnswerTypeException;
-import edu.harvard.h2ms.repository.EventRepository;
 import edu.harvard.h2ms.repository.QuestionRepository;
 import edu.harvard.h2ms.repository.UserRepository;
 import edu.harvard.h2ms.service.utils.H2msRestUtils;
-import java.io.IOException;
-import java.io.StringWriter;
-import java.io.Writer;
+import edu.harvard.h2ms.service.utils.ReportUtils;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 import org.hibernate.Hibernate;
 import org.slf4j.Logger;
@@ -29,10 +24,6 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
 
   final Logger log = LoggerFactory.getLogger(ReportWorkerComplianceWarning.class);
 
-  @Autowired private UserService userService;
-
-  @Autowired private EventRepository eventRepository;
-
   @Autowired private EventService eventService;
 
   @Autowired private QuestionRepository questionRepository;
@@ -42,32 +33,6 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
   @Override
   public String getType() {
     return "complianceWarning";
-  }
-
-  /**
-   * Writes Report using list of list of strings
-   *
-   * @param data [[string,string,],[....]]
-   * @return CSV text
-   */
-  private Writer stringWriterReport(List<List<String>> data) {
-
-    Writer writer = new StringWriter();
-    CSVWriter csvWriter =
-        new CSVWriter(
-            writer,
-            CSVWriter.DEFAULT_SEPARATOR,
-            CSVWriter.DEFAULT_QUOTE_CHARACTER,
-            CSVWriter.DEFAULT_ESCAPE_CHARACTER,
-            CSVWriter.DEFAULT_LINE_END);
-
-    data.forEach((ls) -> csvWriter.writeNext(ls.toArray(new String[0])));
-    try {
-      csvWriter.close();
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-    return writer;
   }
 
   public String createReport() {
@@ -81,9 +46,8 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
 
       try {
         events = eventService.findEventsForCompliance(question);
-        Map<String, Set<Event>> values = new HashMap<>();
 
-        // for all users:
+        // get compliance rates for all users:
         for (User user : userRepository.findAll()) {
           complianceResult.put(
               user,
@@ -98,12 +62,11 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
         allComplianceResult.put(question, complianceResult);
 
       } catch (InvalidAnswerTypeException e) {
-        /// e.printStackTrace();
         log.debug("*********skipping");
       }
     }
-    //    List<List<String>> data = new ArrayList<>();
-    //    Map<Question, Map<User, Double>> allComplianceResult = new HashMap<>();
+
+    // Create data for CSV-like string output
     List<List<String>> data = new ArrayList<List<String>>();
     List<String> row = new ArrayList<String>();
     for (Question question : allComplianceResult.keySet()) {
@@ -119,50 +82,10 @@ public class ReportWorkerComplianceWarning implements ReportWorker {
       }
     }
 
-    Writer writer = stringWriterReport(data);
-
-    try {
-      writer.flush();
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-
-    return writer.toString();
+    return ReportUtils.writeCsvString(data);
   }
 
   public boolean isTriggered() {
-
-    log.info("complianceWarningreport is requested " + questionRepository);
-
-    List<Event> events = new ArrayList<>();
-
-    // for all questions:
-    for (Question question : questionRepository.findAll()) {
-
-      if (question.getAnswerType().equals("boolean")) {
-        Map<User, Double> complianceResult = new HashMap<>();
-        try {
-          events = eventService.findEventsForCompliance(question);
-          // for all users:
-          for (User user : userRepository.findAll()) {
-            complianceResult.put(
-                user,
-                H2msRestUtils.calculateCompliance(
-                    question,
-                    events
-                        .stream()
-                        .filter(event -> event.getSubject().equals(user))
-                        .collect(Collectors.toSet())));
-          }
-
-        } catch (InvalidAnswerTypeException e) {
-          //    			 e.printStackTrace();
-          log.info("****skipping***");
-        }
-      }
-
-      Map<String, Set<Event>> values = new HashMap<>();
-    }
 
     return true;
   }
